add gem pundit, create policies for recources, add authorize method in resources actions

+++ app/controllers/authors_controller.rb
@@ -1,10 +1,14 @@
 class AuthorsController < ApplicationController
+  after_action :verify_authorized
+
   def new
     @author = Author.new
+    authorize @author
   end
 
   def create
     @author = Author.new(author_params)
+    authorize @author
     if @author.save
       flash[:success] = "New author was created"
       redirect_to @author
@@ -15,19 +19,23 @@ class AuthorsController < ApplicationController
 
   def show
     @author = Author.find(params[:id])
+    authorize @author
     @books = @author.books
   end
 
   def index
     @authors = Author.all
+    authorize @authors
   end
 
   def edit
     @author = Author.find(params[:id])
+    authorize @author
   end
 
   def update
     @author = Author.find(params[:id])
+    authorize @author
 
     if @author.update(author_params)
       redirect_to author_path(@author)
@@ -39,6 +47,8 @@ class AuthorsController < ApplicationController
 
   def destroy
     @author = Author.find(params[:id])
+    authorize @author
+
     if @author.destroy
       flash[:success] = "Author deleted"
       redirect_to root_path
