+++ app/controllers/books_controller.rb
@@ -1,18 +1,24 @@
 class BooksController < ApplicationController
+  after_action :verify_authorized
+
   def index
     @books = Book.all
+    authorize @books
   end
 
   def show
     @book = Book.find(params[:id])
+    authorize @book
   end
 
   def edit
     @book = Book.find(params[:id])
+    authorize @book
   end
 
   def update
     @book = Book.find(params[:id])
+    authorize @book
 
     if @book.update(book_params)
       redirect_to book_path(@book)
@@ -24,10 +30,13 @@ class BooksController < ApplicationController
 
   def new
     @book = Book.new
+    authorize @book
+
   end
 
   def create
     @book = Book.new(book_params)
+    authorize @book
 
     if @book.save
       flash[:notice] = "New book was created"
@@ -39,6 +48,8 @@ class BooksController < ApplicationController
 
   def destroy
     @book = Book.find(params[:id])
+    authorize @book
+
     if @book.destroy
       flash[:notice] = "Book deleted"
       redirect_to root_path
