+++ app/controllers/groups_controller.rb
@@ -1,10 +1,15 @@
 class GroupsController < ApplicationController
+  after_action :verify_authorized
+
   def new
     @group = Group.new
+    authorize @group
   end
 
   def create
     @group = Group.new(group_params)
+    authorize @group
+
     if @group.save
       flash[:success] = "New group was created"
       redirect_to @group
@@ -15,15 +20,18 @@ class GroupsController < ApplicationController
 
   def show
     @group = Group.find(params[:id])
+    authorize @group
     @books = @group.books
   end
 
   def edit
     @group = Group.find(params[:id])
+    authorize @group
   end
 
   def  update
     @group = Group.find(params[:id])
+    authorize @group
 
     if @group.update(group_params)
       redirect_to group_path(@group)
@@ -35,10 +43,13 @@ class GroupsController < ApplicationController
 
   def index
     @groups = Group.all
+    authorize @groups
   end
 
   def destroy
     @group = Group.find(params[:id])
+    authorize @group
+
     if @group.destroy
       flash[:success] = "Group deleted"
       redirect_to root_path
