+++ app/controllers/users_controller.rb
@@ -1,14 +1,20 @@
 class UsersController < ApplicationController
+  after_action :verify_authorized
+
   def index
     @users = User.all
+    authorize @users
   end
 
   def show
     @user = User.find(params[:id])
+    authorize @user
   end
 
   def destroy
     @user = User.find(params[:id])
+    authorize @user
+
     if @user.destroy
       flash[:success] = "User deleted"
       redirect_to users_path
@@ -17,10 +23,13 @@ class UsersController < ApplicationController
 
   def edit
     @user = User.find(params[:id])
+    authorize @user
   end
 
   def update
     @user = User.find[params[:id]]
+    authorize @user
+
     if @user.update(user_params)
       redirect_to user_path(admin)
       flash[:notice] = "User permissions successfully updated"
